% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_ini_fxns.R
\name{activate_input_variables}
\alias{activate_input_variables}
\title{Activate variables in the Hector ini files.}
\usage{
activate_input_variables(lines, vars)
}
\arguments{
\item{lines}{character containing the Hector ini input.}

\item{vars}{a vector of the character files.}
}
\value{
the lines of the modified ini file.
}
\description{
Manipulate the ini file by activating variables. This is
useful for setting up the ini files to read in constraints.
}
