% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fxns.R
\name{format_hector_input_table}
\alias{format_hector_input_table}
\title{Format the Hector input into the expected Hector input csv file}
\usage{
format_hector_input_table(x, filename)
}
\arguments{
\item{x}{a data table containing the Hector input information}

\item{filename}{character path for where to save the output.}
}
\value{
path to the csv file formatted as a Hector input table.
}
\description{
Format the Hector input into the expected Hector input csv file
}
