% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_fxns.R
\name{generate_inputs}
\alias{generate_inputs}
\title{Generate emission and concentration constraint csv input tables for Hector.}
\usage{
generate_inputs(scenarios, write_to, years = 1750:2100)
}
\arguments{
\item{scenarios}{a character vector of scenario names.}

\item{write_to}{a character of the full path to the directory location for
the user to write the generated csv files to.}

\item{years}{a vector of the expected output years, default is a sequence from 1750 to 2100.}
}
\value{
the location of the generated csv files.
}
\description{
Generate emission and concentration constraint csv input tables for Hector.
}
