% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_ini_fxns.R
\name{make_new_ini}
\alias{make_new_ini}
\title{Generate a new ini file for a given emissions file}
\usage{
make_new_ini(files)
}
\arguments{
\item{files}{str vector of the new csv hector input files.}
}
\value{
str of the ini files
}
\description{
Generate a new ini file for a given emissions file
}
