% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_ini_fxns.R
\name{replace_csv_string}
\alias{replace_csv_string}
\title{Replace the emissions csv file strings with the path to the new csv table}
\usage{
replace_csv_string(
  ini,
  replacement_path,
  run_name,
  pattern = "=csv:.*TEMPLATE_emissions.csv"
)
}
\arguments{
\item{ini}{lines of a Hector ini file.}

\item{replacement_path}{the file path to the hector input csv table.}

\item{run_name}{character name of run name.}

\item{pattern}{regular expression pattern of the csv paths to replace with \code{replacement_path}
the default is set to replace all emission and concentration constraints.}
}
\value{
lines of a Hector ini file.
}
\description{
Replace the emissions csv file strings with the path to the new csv table
}
