% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{template_ini}
\alias{template_ini}
\title{template_ini}
\format{
A vector
\describe{ A character vector where each element in the vector contains a different
line of hector ini file.
}
}
\usage{
template_ini
}
\description{
A character vector consisting of Hector inputs.
}
\keyword{datasets}
