% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fxns.R
\name{write_hector_csv}
\alias{write_hector_csv}
\title{Save the hector csv files into the proper hector format}
\usage{
write_hector_csv(x, write_to)
}
\arguments{
\item{x}{a datatable emissions or concentration data for a single emissions data frame}

\item{write_to}{str directory to write the hector csv output to}
}
\value{
str file name
}
\description{
Save the hector csv files into the proper hector format
}
